refactor: decompose large UI files into smaller components
This commit improves code organization by breaking down large monolithic
UI files into smaller, more focused components following the Single
Responsibility Principle.

Changes include:

- Refactored activity_page.dart:
  - Created separate components for notification items and groups
  - Added dedicated error and empty state components
  - Created separate list views for grouped and flat notification displays

- Refactored tasks_detail_page.dart:
  - Extracted task header, description, and info components
  - Created dedicated components for reassignment notices
  - Separated progress section and status action components
  - Added a reusable status update dialog

- Refactored login_page.dart:
  - Extracted login form, app logo, and button components
  - Created a dedicated connection status component

This refactoring improves maintainability, makes the code easier to read
and understand, enables better component reuse, and follows the same
structural patterns used elsewhere in the codebase.
