+++ flutter_mobile_app/lib/custom_navigation_bar.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_mobile_app/app_theme.dart';
-import 'package:flutter_mobile_app/pages/login_page.dart';
+import 'package:flutter_mobile_app/pages/login/login_page.dart';
 import 'package:flutter_mobile_app/pages/tasks/tasks_page.dart';
-import 'package:flutter_mobile_app/pages/activity_page.dart';
+import 'package:flutter_mobile_app/pages/activity/activity_page.dart';
 import 'package:flutter_mobile_app/pages/team/team_page.dart';
 import 'package:flutter_mobile_app/pages/profile/profile_page.dart';
 import 'package:flutter_mobile_app/services/api_services.dart';
