+++ flutter_mobile_app/lib/pages/login/login_button.dart
@@ -1,0 +1,35 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_mobile_app/app_theme.dart';
+
+class LoginButton extends StatelessWidget {
+  final bool isLoading;
+  final VoidCallback onPressed;
+
+  const LoginButton({
+    super.key,
+    required this.isLoading,
+    required this.onPressed,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    if (isLoading) {
+      return CircularProgressIndicator(
+        valueColor: AlwaysStoppedAnimation<Color>(AppTheme.primaryColor),
+      );
+    }
+
+    return ElevatedButton(
+      onPressed: onPressed,
+      style: ElevatedButton.styleFrom(
+        minimumSize: Size(double.infinity, 50),
+        backgroundColor: AppTheme.primaryColor,
+        foregroundColor: AppTheme.textOnPrimaryColor,
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(AppTheme.borderRadiusMd),
+        ),
+      ),
+      child: Text('Login', style: AppTheme.buttonTextStyle),
+    );
+  }
+}
